use name similaraity algo
